Add extensible extractor APIs
