+++ src/main/scala/com/linkedin/feathr/offline/source/dataloader/AvroJsonDataLoader.scala
@@ -15,6 +15,7 @@ import org.apache.spark.sql.{DataFrame, Row, SparkSession}
 import java.io.{ByteArrayInputStream, DataInputStream}
 import java.util
 import scala.collection.convert.wrapAll._
+import scala.io.Source
 import scala.reflect.ClassTag
 import scala.util.Try
 
@@ -103,7 +104,7 @@ private[offline] object AvroJsonDataLoader {
     val sc = ss.sparkContext
     require(sc.isLocal)
     require(path.endsWith(".avro.json"))
-    val contents =  scala.io.Source.fromFile(s"src/test/resources/" + path).getLines().mkString
+    val contents = Source.fromResource(path).getLines().mkString
     val jackson = new ObjectMapper(new HoconFactory)
     val tree = jackson.readTree(contents)
     val jsonDataArray = tree.get("data")
